requirements.txt
Pillow==10.4.0
pillow_heif==0.22.0
numpy==2.2.4
scipy==1.15.2
trimesh==4.6.8
